Update initialization and start

+++ cli/main.go
@@ -5,9 +5,11 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
+	"github.com/fatih/color"
 	"github.com/urfave/cli"
 	"io/ioutil"
 	"os"
+	"strings"
 )
 
 func main() {
@@ -21,10 +23,10 @@ func main() {
 			Action:      initAction,
 		},
 		{
-			Name:        "create",
-			Description: "Create new sweady prod cluster",
+			Name:        "start",
+			Description: "Start sweady prod cluster",
 			Category:    "Sweady",
-			Action:      createAction,
+			Action:      startAction,
 		},
 	}
 
@@ -32,41 +34,69 @@ func main() {
 }
 
 func initAction(c *cli.Context) error {
+
+	color.Green("Initialization of Sweady")
+
+	if !askForConfirmation() {
+		color.Yellow("Initialization cancelled")
+		return nil
+	}
+
 	prettyConf, _ := json.Marshal(config.Init())
 	prettyConf, _ = prettyJson(prettyConf)
 
+	color.Green("Add sweady_config.json")
 	err := ioutil.WriteFile("./sweady_config.json", prettyConf, 0644)
-	check(err)
+	if err != nil {
+		return cli.NewExitError(err.Error(), 86)
+	}
 
+	color.Green("Initialization completed")
 	return nil
 }
 
-func createAction(c *cli.Context) error {
+func startAction(c *cli.Context) error {
 
-	fmt.Println("Initialization of template")
-	dat, err := ioutil.ReadFile("./config.dist.json")
+	color.Green("Welcome with Sweady")
+	dat, err := ioutil.ReadFile("./sweady_config.json")
 	if err != nil {
-		fmt.Print(err)
+		return cli.NewExitError(err.Error(), 86)
 	}
+	color.Green("Config found [OK]")
 
 	var i config.ConfValidatorInterface = config.JsonValidator{string(dat)}
 	if i.IsValid() {
-		fmt.Print("Valid")
+		color.Green("Config syntax [OK]")
 	} else {
-		fmt.Print("Non valid")
+		return cli.NewExitError("Config syntax [KO]", 86)
 	}
+
+	color.Green("Creation of environnement file")
+	color.Green("Starting cluster")
 
 	return nil
-}
-
-func check(e error) {
-	if e != nil {
-		panic(e)
-	}
 }
 
 func prettyJson(b []byte) ([]byte, error) {
 	var out bytes.Buffer
 	err := json.Indent(&out, b, "", "  ")
 	return out.Bytes(), err
+}
+
+func askForConfirmation() bool {
+	var s string
+
+	color.Yellow("Are you sure ? (y/N): ")
+	_, err := fmt.Scan(&s)
+	if err != nil {
+		panic(err)
+	}
+
+	s = strings.TrimSpace(s)
+	s = strings.ToLower(s)
+
+	if s == "y" || s == "yes" {
+		return true
+	}
+	return false
 }
